phaco_geocode : Geocodeur pour la Belgique

Cette fonction est la principale du package phacochr. A partir d’une liste d’adresses, elle permet de retrouver leurs coordonnees X-Y.

Usage

phaco_geocode(
  data_to_geocode,
  colonne_rue,
  colonne_num = NULL,
  colonne_code_postal = NULL,
  colonne_num_rue = NULL,
  colonne_num_rue_code_postal = NULL,
  colonne_rue_code_postal = NULL,
  method_stringdist = "lcs",
  corrections_REGEX = TRUE,
  error_max = 4,
  approx_num_max = 50,
  elargissement_com_adj = TRUE,
  lang_encoded = c("FR", "NL", "DE")
)

Arguments

data_to_geocode

Un dataframe avec les adresses a geocoder

colonne_rue

Nom de la colonne avec les rues

colonne_num

Nom de la colonne avec les numéros

colonne_code_postal

Nom de la colonne avec les codes postaux

colonne_num_rue

Nom de la colonne avec numéros et rues ensembles

colonne_num_rue_code_postal

Nom de la colonne avec numéros, rues et code postaux ensembles

colonne_rue_code_postal

Nom de la colonne avec rues et code postaux ensembles

method_stringdist

Méthode pour la jointure inexacte. Par défaut: "lcs". Choix possibles: "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw","soundex".

corrections_REGEX

Correction orthographique. Par défaut: TRUE. Cette option n'est désactivable que si la rue est contenue dans une colonne séparée, et ne contient ni le numéro ni le code postal.

error_max

Nombre maximal d'erreurs entre le nom de la rue a trouver et le nom de la rue dans la base de donnée de référence (BeST). Par défaut: TRUE

approx_num_max

Nombre de numéros d'écart maximum si le numéro n'a pas été trouve. Par défaut: 50

elargissement_com_adj

Élargissement aux communes limitrophes. Par défaut: TRUE

lang_encoded

Langue utilisée pour encoder les noms de rue. Par défaut: c("FR", "NL", "DE")

Examples

x <- data.frame(nom = c(paste0("Observatoire de la Sant","\u00e9"," et du Social"), "ULB"),
rue = c("rue Belliard","avenue Antoine Depage"),
num = c("71", "30"),
code_postal = c("1040","1000"))

result <- phaco_geocode(data_to_geocode = x,
colonne_rue = "rue",
colonne_num = "num",
colonne_code_postal = "code_postal")
#> --- PhacochR ---
#> -- Formatage des données
#> → Champs introduits : numéro, rue et code postal séparés
#> ℹ Région(s) détectée(s) : Bruxelles
#> ⧗ Correction orthographique des adresses✔ Correction orthographique des adresses
#> -- Géocodage
#> ⧗ Paramétrage pour utiliser 7 coeurs de l'ordinateur✔ Paramétrage pour utiliser 7 coeurs de l'ordinateur
#> ⧗ Détection des rues (matching inexact avec fuzzyjoin)✔ Détection des rues (matching inexact avec fuzzyjoin)
#> ⧗ Élargissement pour les rues non trouvées aux communes adjacentes✔ Élargissement pour les rues non trouvées aux communes adjacentes
#> ⧗ Chargement du fichier openaddress✔ Chargement du fichier openaddress 
#> ⧗ Jointure avec les coordonnées X-Y✔ Jointure avec les coordonnées X-Y
#> ⧗ Approximation à + ou - 100 numéros pour les adresses non localisées✔ Approximation à + ou - 100 numéros pour les adresses non localisées
#> -- Résultats
#> ⧗ Création du fichier final et formatage des tables de vérification✔ Création du fichier final et formatage des tables de vérification
#> ✔ Géocodage terminé
#> ℹ Statistiques concernant le géocodage:
#> 
#> |Region    |  n| Rue detect.(%) | Approx. num(n) | Elarg. com.(n) | Abrev. noms(n) | Geocode(%) |
#> |:---------|--:|:--------------:|:--------------:|:--------------:|:--------------:|:----------:|
#> |Bruxelles |  2|      100       |       0        |       0        |       0        |    100     |
#> |Total     |  2|      100       |       0        |       0        |       0        |    100     |
#> 
#> ℹ Temps de calcul total : 8.9 s
#>              
#> /!\ Toutes les adresses n'ont pas été trouvées avec certitude /!\
#> - check 'dist_fuzzy' pour les erreurs de reconnaissance des rues
#> - check 'approx_num' pour les approximations de numéro
#> - check 'type_geocoding' pour l'éargissement aux communes adjacentes
#> - check 'nom_propre_abv' pour les abréviations de noms propres
#>              
#> -- Plus de résultats:
#> → Tableau synthétique : $summary
#> → Données géocodées : $data_geocoded
#> → Données géocodées en format sf : $data_geocoded_sf